requirements.txt
python-dotenv==1.0.1
sockets==1.0.0
